feat!: 18.0.0 - fish catch-all, qstlist flat quest entry, NsTeST unknown fields
BREAKING CHANGES:
- FishPacket: reduced to Type + string Body catch-all (structure beyond Type is
  not reverse-engineered yet; preserves round-trip without lying about fields)
- QuestSubPacket: flat 16-field tuple per vanosilla layout
  <slot>.<questId>.<questLineId>.<type>.<obj1Cur>.<obj1Req>.<status>.
  <obj2Cur>.<obj2Req>...<obj5Cur>.<obj5Req>.<showInfo>.
  Removes ObjectiveCount / InfoId / GoalType / QuestObjectiveSubPackets /
  ShowDialog; deletes QuestObjectiveSubPacket
- NsTeSTPacket: adds 6 int Unknown fields (observed wire `1 0 11 4 13 3`)
  between Unknown and ServerCharacters; indices shifted accordingly
- NsTeStSubPacket.Port widened to int? so the -1:-1:-1 trailing sentinel
  round-trips without ushort overflow

+++ src/NosCore.Packets/ServerPackets/Login/NsTeSTSubPacket.cs
@@ -16,7 +16,7 @@ namespace NosCore.Packets.ServerPackets.Login
         public string? Host { get; set; }
 
         [PacketIndex(1, SpecialSeparator = ":")]
-        public ushort? Port { get; set; }
+        public int? Port { get; set; }  // int so the -1:-1:-1:10000.10000.1 sentinel round-trips without ushort overflow
 
         [PacketIndex(2, SpecialSeparator = ":")]
         public int? Color { get; set; }
